feat(client-generator-js): rename `/wasm` to `/edge`
Rename `/wasm` entrypoint to `/edge` to ensure smooth migration for
Accelerate users.

Closes: https://linear.app/prisma-company/issue/TML-1566/rename-wasm-entrypoint-to-edge-in-prismaclient-and-old-generator
