+++ packages/bundle-size/da-workers-d1/index.js
@@ -1,6 +1,6 @@
 import { PrismaD1 } from '@prisma/adapter-d1'
 
-import { PrismaClient } from './client/wasm'
+import { PrismaClient } from './client/edge'
 
 export default {
   async fetch(request, env) {
@@ -10,7 +10,6 @@ export default {
     const users = await prisma.user.findMany()
     const result = JSON.stringify(users)
 
-    // eslint-disable-next-line no-undef
     return new Response(result)
   },
 }
