+++ packages/bundle-size/da-workers-mariadb/index.js
@@ -1,6 +1,6 @@
 import { PrismaMariaDb } from '@prisma/adapter-mariadb'
 
-import { PrismaClient } from './client/wasm'
+import { PrismaClient } from './client/edge'
 
 export default {
   async fetch(request, env) {
@@ -12,7 +12,6 @@ export default {
     const users = await prisma.user.findMany()
     const result = JSON.stringify(users)
 
-    // eslint-disable-next-line no-undef
     return new Response(result)
   },
 }
