+++ packages/bundle-size/da-workers-mssql/index.js
@@ -1,6 +1,6 @@
 import { PrismaMssql } from '@prisma/adapter-mssql'
 
-import { PrismaClient } from './client/wasm'
+import { PrismaClient } from './client/edge'
 
 export default {
   async fetch(request, env) {
@@ -12,7 +12,6 @@ export default {
     const users = await prisma.user.findMany()
     const result = JSON.stringify(users)
 
-    // eslint-disable-next-line no-undef
     return new Response(result)
   },
 }
