+++ packages/client/edge.d.ts
@@ -1,0 +1,1 @@
+export * from '.prisma/client/edge'
