+++ packages/client/edge.js
@@ -1,4 +1,4 @@
 module.exports = {
   // https://github.com/prisma/prisma/pull/12907
-  ...require('.prisma/client/wasm'),
+  ...require('.prisma/client/edge'),
 }
